# docs/v2/api-spec.yaml
openapi: 3.0.3
info:
  title: Timesheet API
  version: 0.2.0
  description: |
    API for the Timesheet application v2.

    This is the source of truth for the API contract. All clients (Web, MCP, CLI)
    consume this spec. The Go service is generated from this spec using oapi-codegen.
  x-mcp:
    name: timesheet
    version: 1.0.0
    instructions: |
      You are an AI assistant helping manage a timesheet application.

      The user tracks their time across different projects. Calendar events are synced from
      Google Calendar and need to be classified (assigned to projects or marked as skipped).

      IMPORTANT: Before using search_events, create_rule, or preview_rule tools, first read the
      timesheet://docs/query-syntax resource to understand the query language.

      When helping the user:
      1. Read timesheet://docs/query-syntax to learn the search syntax
      2. List projects to understand available classification targets
      3. Search for pending events to see what needs attention
      4. Use preview_rule to test classification patterns
      5. Create rules or use bulk_classify to classify events
    resources:
      - uri: "timesheet://docs/query-syntax"
        name: "Query Syntax Reference"
        description: "Complete reference for the Gmail-style query syntax used to search events and create classification rules"
        mimeType: "text/markdown"
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/auth/google/callback:
    get:
      operationId: googleCallback
      tags: [calendars]
      summary: Handle Google OAuth callback
      security:
        - bearerAuth: []
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: Authorization code from Google
        - name: state
          in: query
          required: true
          schema:
            type: string
          description: State parameter for CSRF protection
      responses:
        '201':
          description: Calendar connected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarConnection'
        '400':
          description: Invalid request (missing code/state)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login/signup, or API key (format ts_xxx)
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true

    # Calendar schemas
...
    CalendarConnection:
      type: object
      required: [id, user_id, provider, created_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        provider:
          type: string
          enum: [google]
        last_synced_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...